+++ components/gnutella-core/src/main/java/com/limegroup/gnutella/udpconnect/WriteRegulator.java
@@ -274,7 +274,7 @@ public class WriteRegulator {
 		if ( gettingSlow > 0 )
 			gettingSlow--;
 
-        return (long) sleepTime;
+        return sleepTime;
         //------------- Sleep ------------------------
     }
